backend/requirements.txt
fastapi 
uvicorn 
python-arango
boto3
python-multipart
python-dotenv